# readium-api/utils/swagger/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Readium API",
    "description": "Handsome backend devs"
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    }
  ],
...
  "paths": {
...
    "/auth/": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User Login",
        "description": "",
        "parameters": [],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUser"
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "User Logout",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/auth/logout-all": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "User Logout from all devices",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User Register by email",
        "description": "",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Account created"
          },
          "400": {
            "description": "Email has already been used or fields have errors"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUser"
              }
            }
          }
        }
      }
    },
    "/auth/confirm": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "User CONFIRM their account",
        "description": "",
        "parameters": [
          {
            "name": "iv",
            "in": "query",
            "type": "string",
            "required": true
          },
          {
            "name": "id",
            "in": "query",
            "type": "string",
            "description": "hashed user id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Activate successfully"
          },
          "400": {
            "description": "Account has already activated"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
    "/auth/reset": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Send reset password link to authenticated users",
        "description": "",
        "parameters": [],
        "responses": {},
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Users reset password",
        "description": "",
        "parameters": [
          {
            "name": "iv",
            "in": "query",
            "type": "string",
            "required": true
          },
          {
            "name": "id",
            "in": "query",
            "type": "string",
            "description": "hashed user id",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "User not found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPassword"
              }
            }
          }
        }
      }
    },
    "/auth/forget": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Users forget password",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Request change password successfully "
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Email"
              }
            }
          }
        }
      }
    },
    "/auth/change-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "User change password",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Saving changed password successfully"
          },
          "400": {
            "description": "Typed in wrong password"
          },
          "500": {
            "description": "Saving user to mongodb has some errors"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePassword"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RegisterUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "example": "testing"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "password": {
            "type": "string",
            "example": "testing"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ChangePassword": {
        "type": "object",
        "properties": {
          "oldPassword": {
            "type": "string",
            "example": "testing"
          },
          "password": {
            "type": "string",
            "example": "testing123"
          }
        },
        "required": [
          "oldPassword",
          "password"
        ]
      },
      "ResetPassword": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "example": "testing123"
          }
        },
        "required": [
          "password"
        ]
      },
...
      "Email": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}